Add reservation app
Create Appointment, TypeOfJob, Schedule models

At the time of creating this PR we still don't have the Client and Professional apps.

The model has foreign keys from client and professional which have not been merged yet, so for now it was added within a comment.

When we add the Client and Professional apps, I'll make the necessary changes in the models.py.