Change logic and tests of the types of jobs of a professional after login PR merge
I adjusted the feature design to be part of the system design.

I added in the tests the option to check whether a user is connected or not.

I removed the typeOfJob_confirm_delete.html and when the user clicks on the delete button it does not move him to another page.
Instead, it simply removes the object and stays on the same page.