# This is a combination of 8 commits. # This is the 1st commit message:
Making profile adjustments

Profile sync to the system theme.

Users are required to login to the site.

URL for a logged in user is http://127.0.0.1:8000/profile/ - no matter if it is a Client or a Professional.

Signed-off-by: TalReinfeld <talreinfeld1@gmail.com>

# This is the commit message #2:

Making profile adjustments

Profile sync to the system theme.

Users are required to login to the site.

URL for a logged in user is http://127.0.0.1:8000/profile/ - no matter if it is a Client or a Professional.

Signed-off-by: TalReinfeld <talreinfeld1@gmail.com>

# This is the commit message #3:

Making profile adjustments

Profile sync to the system theme.

Users are required to login to the site.

URL for a logged in user is http://127.0.0.1:8000/profile/ - no matter if it is a Client or a Professional.

Signed-off-by: TalReinfeld <talreinfeld1@gmail.com>

# This is the commit message #4:

Making profile adjustments

Profile sync to the system theme.

Users are required to login to the site.

URL for a logged in user is http://127.0.0.1:8000/profile/ - no matter if it is a Client or a Professional.

Signed-off-by: TalReinfeld <talreinfeld1@gmail.com>

# This is the commit message #5:

Making profile adjustments

Profile sync to the system theme.

Users are required to login to the site.

URL for a logged in user is http://127.0.0.1:8000/profile/ - no matter if it is a Client or a Professional.

Signed-off-by: TalReinfeld <talreinfeld1@gmail.com>

# This is the commit message #6:

# This is a combination of 16 commits.
parent b38426ab7c107510176408acc81f9c970a035123
author TalReinfeld <talreinfeld1@gmail.com> 1685636677 +0300
committer TalReinfeld <talreinfeld1@gmail.com> 1685636677 +0300

Making profile adjustments

Profile sync to the system theme.

Users are required to login to the site.

URL for a logged in user is http://127.0.0.1:8000/profile/ - no matter if it is a Client or a Professional.

Signed-off-by: TalReinfeld <talreinfeld1@gmail.com>

# This is the commit message #14:

Modified code that broke the tests

# This is the commit message #16:

 I fixed the file account/tests/test_login.py

 1.I added a check that the login.html file is used
 2.Add Login Form validation

# This is the commit message #7:

Making profile adjustments

Profile sync to the system theme.

Users are required to login to the site.

URL for a logged in user is http://127.0.0.1:8000/profile/ - no matter if it is a Client or a Professional.

Signed-off-by: TalReinfeld <talreinfeld1@gmail.com>

# This is the commit message #8:

Making profile adjustments

Profile sync to the system theme.

Users are required to login to the site.

URL for a logged in user is http://127.0.0.1:8000/profile/ - no matter if it is a Client or a Professional.

Signed-off-by: TalReinfeld <talreinfeld1@gmail.com>